pyproject.toml
[project]
name = "abitur"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ruff",
    "aiogram",
    "python-dotenv",
    "httpx",
    "bs4>=0.0.2",
    "requests>=2.32.5",
    "pyyaml>=6.0.3",
]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
fix = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
